Minor change in Name
This is not a big change by any means but the word Widget sounds more native than View when we are talking about Flutter Apps. 
I also suspect that you might have named it View with specific reason instead of Widgets intentionally. If so can I know the reason please?